refactor(insee): extract insee connector to individual pkg
