+++ .changeset/hip-houses-confess.md
@@ -1,0 +1,7 @@
+---
+"@gouvfr-lasuite/proconnect.insee": minor
+---
+
+♻️ Prélevement d'un partie du connecteur INSEE
+
+Dans le cadres la migration du script d'import de comptes coop, une partie de l'API INSEE est déplacées dans le package `@gouvfr-lasuite/proconnect.insee` pour permettre leur réutilisation dans Hyyypertool.
