+++ .github/workflows/test.yml
@@ -19,7 +19,9 @@ jobs:
         with:
           cache: "npm"
           node-version-file: package.json
-      - run: CYPRESS_INSTALL_BINARY=0 npm ci --include=dev
+      - run: npm ci
+        env:
+          CYPRESS_INSTALL_BINARY: 0
       - run: npm run build:workspaces
       - run: npm run test:lint
       - run: npm run test:workspaces
