+++ Dockerfile
@@ -5,8 +5,9 @@ WORKDIR /app
 FROM base AS prod-deps
 RUN --mount=type=bind,source=package.json,target=package.json \
   --mount=type=bind,source=package-lock.json,target=package-lock.json \
-  --mount=type=bind,source=packages/email/package.json,target=packages/email/package.json \
   --mount=type=bind,source=packages/core/package.json,target=packages/core/package.json \
+  --mount=type=bind,source=packages/email/package.json,target=packages/email/package.json \
+  --mount=type=bind,source=packages/insee/package.json,target=packages/insee/package.json \
   --mount=type=cache,target=/root/.npm \
   npm ci --omit=dev
 
@@ -14,8 +15,9 @@ FROM base AS build
 ENV CYPRESS_INSTALL_BINARY=0
 RUN --mount=type=bind,source=package.json,target=package.json \
   --mount=type=bind,source=package-lock.json,target=package-lock.json \
-  --mount=type=bind,source=packages/email/package.json,target=packages/email/package.json \
   --mount=type=bind,source=packages/core/package.json,target=packages/core/package.json \
+  --mount=type=bind,source=packages/email/package.json,target=packages/email/package.json \
+  --mount=type=bind,source=packages/insee/package.json,target=packages/insee/package.json \
   --mount=type=cache,target=/root/.npm \
   npm ci
 COPY tsconfig.json vite.config.mjs ./
