+++ package.json
@@ -38,7 +38,8 @@
     "watch:js": "tsc --watch --preserveWatchOutput",
     "watch:node": "tsx --watch src/index.ts",
     "watch:workspaces:core": "npm run dev --if-present --workspace=@gouvfr-lasuite/proconnect.core",
-    "watch:workspaces:email": "npm run dev --if-present --workspace=@gouvfr-lasuite/proconnect.email"
+    "watch:workspaces:email": "npm run dev --if-present --workspace=@gouvfr-lasuite/proconnect.email",
+    "watch:workspaces:insee": "npm run dev --if-present --workspace=@gouvfr-lasuite/proconnect.insee"
   },
   "prettier": {
     "plugins": [
@@ -51,6 +52,7 @@
     "@gouvfr-lasuite/crisp": "https://github.com/douglasduteil/crisp/releases/download/v1.6.1/douglasduteil-crisp-1.6.1.tgz",
     "@gouvfr-lasuite/proconnect.core": "workspace:*",
     "@gouvfr-lasuite/proconnect.email": "workspace:*",
+    "@gouvfr-lasuite/proconnect.insee": "workspace:*",
     "@gouvfr/dsfr": "^1.12.1",
     "@kitajs/html": "^4.2.4",
     "@kitajs/ts-html-plugin": "^4.1.0",
@@ -130,7 +132,7 @@
     "cypress-axe": "^1.5.0",
     "cypress-maildev": "^1.3.2",
     "mocha": "^11.0.1",
-    "nock": "^13.5.4",
+    "nock": "^13.5.6",
     "prettier": "^3.4.2",
     "prettier-plugin-organize-imports": "^4.1.0"
   },
