+++ packages/insee/src/data/index.ts
@@ -1,0 +1,6 @@
+//
+
+export * from "./categories-juridiques.js";
+export * from "./codes-effectifs.js";
+export * from "./codes-naf.js";
+export * from "./codes-voie.js";
