+++ packages/insee/src/errors/index.ts
@@ -1,0 +1,8 @@
+//
+
+export class InseeNotFoundError extends Error {}
+
+export class InvalidSiretError extends Error {}
+export class InvalidSirenError extends Error {}
+
+export class InseeConnectionError extends Error {}
