+++ packages/insee/src/formatters/enseigne.ts
@@ -1,0 +1,8 @@
+//
+
+import { capitalize, isEmpty } from "lodash-es";
+
+//
+
+export const formatEnseigne = (...args: (string | null)[]) =>
+  capitalize(args.filter((e) => !isEmpty(e)).join(" ")) || "";
