+++ packages/insee/src/formatters/libelle-from-categories-juridiques.ts
@@ -1,0 +1,9 @@
+//
+
+import { categoriesJuridiques, type CategoriesJuridique } from "#src/data";
+
+//
+
+export const libelleFromCategoriesJuridiques = (
+  categorie: CategoriesJuridique,
+) => categoriesJuridiques[categorie] || null;
