+++ packages/insee/src/types/index.ts
@@ -1,0 +1,4 @@
+//
+
+export * from "./establishment.js";
+export * from "./tranche-effectifs.js";
