+++ packages/insee/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "declaration": true,
+    "declarationMap": true,
+    "outDir": "./dist",
+    "rootDir": "src",
+    "types": ["node"],
+    "module": "NodeNext",
+    "moduleResolution": "nodenext",
+    "verbatimModuleSyntax": true,
+    "paths": {
+      "#src/*": ["./src/*"]
+    }
+  },
+  "extends": "@tsconfig/node22/tsconfig.json",
+  "references": []
+}
