+++ scripts/import-domains.ts
@@ -9,10 +9,7 @@ import fs from "fs";
 import { isEmpty, some, toInteger } from "lodash-es";
 import { z } from "zod";
 import { InseeNotFoundError } from "../src/config/errors";
-import {
-  getInseeAccessToken,
-  getOrganizationInfo,
-} from "../src/connectors/api-sirene";
+import { getOrganizationInfo } from "../src/connectors/api-sirene";
 import {
   addDomain,
   findEmailDomainsByOrganizationId,
@@ -27,6 +24,7 @@ import {
   startDurationMesure,
   throttleApiCall,
 } from "../src/services/script-helpers";
+import type { Organization } from "../src/types/organization";
 
 const { INPUT_FILE, OUTPUT_FILE } = z
   .object({
@@ -44,8 +42,6 @@ const rateInMsFromArgs = toInteger(process.argv[2]);
 const maxInseeCallRateInMs = rateInMsFromArgs !== 0 ? rateInMsFromArgs : 125;
 
 (async () => {
-  const access_token = await getInseeAccessToken();
-
   const readStream = fs.createReadStream(INPUT_FILE); // readStream is a read-only stream wit raw text content of the CSV file
   const writeStream = fs.createWriteStream(OUTPUT_FILE); // writeStream is a write-only stream to write on the disk
 
@@ -131,10 +127,7 @@ const maxInseeCallRateInMs = rateInMsFromArgs !== 0 ? rateInMsFromArgs : 125;
           const start = startDurationMesure();
           try {
             // 2. get organizationInfo
-            const organizationInfo = await getOrganizationInfo(
-              siret,
-              access_token,
-            );
+            const organizationInfo = await getOrganizationInfo(siret);
             await throttleApiCall(start, maxInseeCallRateInMs);
 
             // 3. check organization status
