+++ src/connectors/api-sirene/index.ts
@@ -1,4 +1,18 @@
-import axios, { AxiosError, type AxiosResponse } from "axios";
+import {
+  findBySirenFactory,
+  findBySiretFactory,
+  getInseeAccessTokenFactory,
+} from "@gouvfr-lasuite/proconnect.insee/api";
+import {
+  formatAdresseEtablissement,
+  formatEnseigne,
+  formatNomComplet,
+  libelleFromCategoriesJuridiques,
+  libelleFromCodeEffectif,
+  libelleFromCodeNaf,
+} from "@gouvfr-lasuite/proconnect.insee/formatters";
+import type { InseeEtablissement } from "@gouvfr-lasuite/proconnect.insee/types";
+import { AxiosError } from "axios";
 import { cloneDeep, set } from "lodash-es";
 import {
   HTTP_CLIENT_TIMEOUT,
@@ -10,174 +24,11 @@ import {
   InseeNotFoundError,
   InvalidSiretError,
 } from "../../config/errors";
-import {
-  formatAdresseEtablissement,
-  formatEnseigne,
-  formatNomComplet,
-  libelleFromCategoriesJuridiques,
-  libelleFromCodeEffectif,
-  libelleFromCodeNaf,
-} from "./formatters";
-
-type InseeEtablissement = {
-  // ex: '217400563'
-  siren: string;
-  // ex: '00011'
-  nic: string;
-  // ex: '21740056300011'
-  siret: string;
-  // ex: 'O'
-  statutDiffusionEtablissement: "O" | "P" | "N";
-  // ex: '1983-03-01'
-  dateCreationEtablissement: string;
-  // ex: '32'
-  trancheEffectifsEtablissement: TrancheEffectifs;
-  // ex: '2020'
-  anneeEffectifsEtablissement: string;
-  activitePrincipaleRegistreMetiersEtablissement: string | null;
-  // ex: '2022-08-29T09:08:45'
-  dateDernierTraitementEtablissement: string;
-  // ex: true
-  etablissementSiege: boolean;
-  // ex: 4
-  nombrePeriodesEtablissement: number;
-  uniteLegale: {
-    // ex: 'A'
-    etatAdministratifUniteLegale: string;
-    // ex: 'O'
-    statutDiffusionUniteLegale: "O" | "P" | "N";
-    // ex: '1982-01-01'
-    dateCreationUniteLegale: string;
-    // ex: '7210'
-    categorieJuridiqueUniteLegale: string;
-    // ex: 'COMMUNE DE CHAMONIX MONT BLANC'
-    denominationUniteLegale: string;
-    sigleUniteLegale: string | null;
-    denominationUsuelle1UniteLegale: string | null;
-    denominationUsuelle2UniteLegale: string | null;
-    denominationUsuelle3UniteLegale: string | null;
-    sexeUniteLegale: string | null;
-    nomUniteLegale: string | null;
-    nomUsageUniteLegale: string | null;
-    prenom1UniteLegale: string | null;
-    prenom2UniteLegale: string | null;
-    prenom3UniteLegale: string | null;
-    prenom4UniteLegale: string | null;
-    prenomUsuelUniteLegale: string | null;
-    pseudonymeUniteLegale: string | null;
-    // ex: '84.11Z'
-    activitePrincipaleUniteLegale: string;
-    nomenclatureActivitePrincipaleUniteLegale: "NAFRev2";
-    identifiantAssociationUniteLegale: string | null;
-    // ex: 'N'
-    economieSocialeSolidaireUniteLegale: string;
-    // ex: 'N'
-    societeMissionUniteLegale: string;
-    // ex: 'O'
-    caractereEmployeurUniteLegale: string;
-    // ex: '32'
-    trancheEffectifsUniteLegale: TrancheEffectifs;
-    // ex: '2020'
-    anneeEffectifsUniteLegale: string;
-    // ex: '00011'
-    nicSiegeUniteLegale: string;
-    // ex: '2023-03-01T20:13:11'
-    dateDernierTraitementUniteLegale: string;
-    // ex: 'ETI'
-    categorieEntreprise: string;
-    // ex: '2020'
-    anneeCategorieEntreprise: string;
-  };
-  adresseEtablissement: {
-    complementAdresseEtablissement: string | null;
-    // ex: '38'
-    numeroVoieEtablissement: string;
-    indiceRepetitionEtablissement: string | null;
-    // ex: 'PL'
-    typeVoieEtablissement: string;
-    // ex: 'DE L EGLISE'
-    libelleVoieEtablissement: string;
-    // ex: '74400'
-    codePostalEtablissement: string;
-    // ex: 'CHAMONIX-MONT-BLANC'
-    libelleCommuneEtablissement: string;
-    libelleCommuneEtrangerEtablissement: string | null;
-    distributionSpecialeEtablissement: string | null;
-    // ex: '74056'
-    codeCommuneEtablissement: string;
-    codeCedexEtablissement: string | null;
-    libelleCedexEtablissement: string | null;
-    codePaysEtrangerEtablissement: string | null;
-    libellePaysEtrangerEtablissement: string | null;
-  };
-  adresse2Etablissement: {
-    complementAdresse2Etablissement: null;
-    numeroVoie2Etablissement: null;
-    indiceRepetition2Etablissement: null;
-    typeVoie2Etablissement: null;
-    libelleVoie2Etablissement: null;
-    codePostal2Etablissement: null;
-    libelleCommune2Etablissement: null;
-    libelleCommuneEtranger2Etablissement: null;
-    distributionSpeciale2Etablissement: null;
-    codeCommune2Etablissement: null;
-    codeCedex2Etablissement: null;
-    libelleCedex2Etablissement: null;
-    codePaysEtranger2Etablissement: null;
-    libellePaysEtranger2Etablissement: null;
-  };
-  periodesEtablissement: {
-    dateFin: string | null;
-    // ex: '2008-01-01'
-    dateDebut: string;
-    // ex: 'A'
-    etatAdministratifEtablissement: string;
-    // ex: false
-    changementEtatAdministratifEtablissement: boolean;
-    // ex: 'MAIRIE CHAMONIX - ARGENTIERE'
-    enseigne1Etablissement: string;
-    enseigne2Etablissement: null;
-    enseigne3Etablissement: null;
-    // ex: false
-    changementEnseigneEtablissement: boolean;
-    denominationUsuelleEtablissement: null;
-    // ex: false
-    changementDenominationUsuelleEtablissement: boolean;
-    // ex: '84.11Z'
-    activitePrincipaleEtablissement: string;
-    nomenclatureActivitePrincipaleEtablissement: "NAFRev2";
-    // ex: true
-    changementActivitePrincipaleEtablissement: boolean;
-    // ex: 'O'
-    caractereEmployeurEtablissement: string;
-    // ex: false
-    changementCaractereEmployeurEtablissement: boolean;
-  }[];
-};
-
-type EtablissementSearchBySiretResponse = {
-  etablissement: InseeEtablissement;
-};
-
-type EtablissementSearchResponse = {
-  header: {
-    total: number;
-    debut: number;
-    nombre: number;
-  };
-  etablissements: InseeEtablissement[];
-};
-
-type GetTokenReponse = {
-  access_token: string;
-  scope: "am_application_scope default";
-  token_type: "Bearer";
-  expires_in: number;
-};
+import type { OrganizationInfo } from "../../types/organization-info";
 
 const hideNonDiffusibleData = (
-  etablissement: EtablissementSearchBySiretResponse["etablissement"],
-): EtablissementSearchBySiretResponse["etablissement"] => {
+  etablissement: InseeEtablissement,
+): InseeEtablissement => {
   const hiddenEtablissement = cloneDeep(etablissement);
   set(hiddenEtablissement, "uniteLegale.denominationUniteLegale", null);
   set(hiddenEtablissement, "uniteLegale.sigleUniteLegale", null);
@@ -264,60 +115,40 @@ const hideNonDiffusibleData = (
   return hiddenEtablissement;
 };
 
-export const getInseeAccessToken = async () => {
-  const {
-    data: { access_token },
-  }: AxiosResponse<GetTokenReponse> = await axios.post(
-    "https://api.insee.fr/token",
-    "grant_type=client_credentials",
-    {
-      headers: { "Content-Type": "application/x-www-form-urlencoded" },
-      auth: {
-        username: INSEE_CONSUMER_KEY!,
-        password: INSEE_CONSUMER_SECRET!,
-      },
-      timeout: HTTP_CLIENT_TIMEOUT,
-    },
-  );
+export const getInseeAccessToken = getInseeAccessTokenFactory(
+  {
+    consumerKey: INSEE_CONSUMER_KEY,
+    consumerSecret: INSEE_CONSUMER_SECRET,
+  },
+  {
+    timeout: HTTP_CLIENT_TIMEOUT,
+  },
+);
 
-  return access_token;
-};
+const findBySiret = findBySiretFactory({
+  getInseeAccessToken,
+  config: {
+    timeout: HTTP_CLIENT_TIMEOUT,
+  },
+});
+
+const findBySiren = findBySirenFactory({
+  getInseeAccessToken,
+  config: {
+    timeout: HTTP_CLIENT_TIMEOUT,
+  },
+});
 
 export const getOrganizationInfo = async (
   siretOrSiren: string,
-  provided_access_token?: string,
 ): Promise<OrganizationInfo> => {
   try {
-    let access_token = provided_access_token;
-    if (!access_token) {
-      access_token = await getInseeAccessToken();
-    }
-
     let etablissement: InseeEtablissement;
 
     if (siretOrSiren.match(/^\d{14}$/)) {
-      let { data }: AxiosResponse<EtablissementSearchBySiretResponse> =
-        await axios.get(
-          `https://api.insee.fr/entreprises/sirene/siret/${siretOrSiren}`,
-          {
-            headers: { Authorization: `Bearer ${access_token}` },
-            timeout: HTTP_CLIENT_TIMEOUT,
-          },
-        );
-
-      etablissement = data.etablissement;
+      etablissement = await findBySiret(siretOrSiren);
     } else if (siretOrSiren.match(/^\d{9}$/)) {
-      // siretOrSiren is a siren, we fetch the data of the siege social
-      let { data }: AxiosResponse<EtablissementSearchResponse> =
-        await axios.get(
-          `https://api.insee.fr/entreprises/sirene/siret?q=siren:${siretOrSiren} AND etablissementSiege:true`,
-          {
-            headers: { Authorization: `Bearer ${access_token}` },
-            timeout: HTTP_CLIENT_TIMEOUT,
-          },
-        );
-
-      etablissement = data.etablissements[0];
+      etablissement = await findBySiren(siretOrSiren);
     } else {
       throw new InvalidSiretError();
     }
@@ -388,10 +219,11 @@ export const getOrganizationInfo = async (
       enseigne,
       trancheEffectifs: trancheEffectifsEtablissement,
       trancheEffectifsUniteLegale,
-      libelleTrancheEffectif: libelleFromCodeEffectif(
-        trancheEffectifsEtablissement,
-        anneeEffectifsEtablissement,
-      ),
+      libelleTrancheEffectif:
+        libelleFromCodeEffectif(
+          trancheEffectifsEtablissement,
+          anneeEffectifsEtablissement,
+        ) ?? "",
       etatAdministratif: etatAdministratifEtablissement,
       estActive: etatAdministratifEtablissement === "A",
       statutDiffusion: statutDiffusionEtablissement,
@@ -403,10 +235,9 @@ export const getOrganizationInfo = async (
       libelleActivitePrincipale: libelleFromCodeNaf(
         activitePrincipaleEtablissement,
       ),
-      categorieJuridique: categorieJuridiqueUniteLegale,
-      libelleCategorieJuridique: libelleFromCategoriesJuridiques(
-        categorieJuridiqueUniteLegale,
-      ),
+      categorieJuridique: String(categorieJuridiqueUniteLegale),
+      libelleCategorieJuridique:
+        libelleFromCategoriesJuridiques(categorieJuridiqueUniteLegale) ?? "",
     };
   } catch (e) {
     if (
