+++ src/repositories/organization/getters.ts
@@ -1,5 +1,6 @@
 import type { QueryResult } from "pg";
 import { getDatabaseConnection } from "../../connectors/postgres";
+import type { Organization } from "../../types/organization";
 
 export const findById = async (id: number) => {
   const connection = getDatabaseConnection();
